Changed rating to FloatField in UserSerializer

+++ app_run/serializers.py
@@ -9,7 +9,7 @@ class UserSerializer(serializers.ModelSerializer):
     # Если не поставить read_only=True, DRF подумает, что это поле должно приходить в теле запроса при
     # создании/обновлении — и может вызвать ошибку вроде: "runs_finished" is a required field.
     runs_finished = serializers.IntegerField(read_only=True)
-    rating = serializers.IntegerField(read_only=True)
+    rating = serializers.FloatField(read_only=True)
 
     class Meta:
         model = User
